vanilla_gan_mnist/requirements.txt
Click==7.0
numpy==1.15.4
Pillow==5.4.0
protobuf==3.6.1
PyYAML==3.13
six==1.12.0
tensorboardX==1.4
torch==0.4.1
torchvision==0.2.1
tqdm==4.26.0